hashchk_parser.py renamed to hashchk_terminal.py. Functions related to digest generation, comparison, etc... limited to hashchk only now.  Anything related to terminal use will be limited to hashchk_terminal.py
