+++ src/hash_check/hashchk.py
@@ -5,11 +5,7 @@ from six.moves import range
 
 import hashlib
 import os
-import sys
 import hmac  # Python 2.7 and 3.3+
-
-# from hashchk_parser.py
-from hashchk_parser import (HashChkParser, Output)
 
 
 # TODO: Implement hash method choices
@@ -34,13 +30,13 @@ class HashCheck(object):
     def generate_digest(filename):
         """Returns hexadecimal digest generated from filename"""
 
-        buffer_size = 65536 # Buffer used to cut down on memory for large files.
+        buffer_size = 65536  # Buffer used to cut down on memory for large files.
         blocks = (os.path.getsize(filename) // buffer_size) + 1
 
         hash_digest = hashlib.sha256()
 
         with open(filename, 'rb') as f:
-            #generator expression used for reading file in chunks
+            # generator expression used for reading file in chunks
             generator = (f.read(buffer_size) for _ in range(blocks))
             for data in generator:
                 hash_digest.update(data)
@@ -52,29 +48,3 @@ class HashCheck(object):
         """Returns True if digest_1 == digest_2"""
 
         return hmac.compare_digest(digest_1, digest_2)
-
-
-def _compare_verify_digests(verify_args):
-    """Takes in parsed arguments from HashChkParserverify subparser and prints
-    out comparison results."""
-
-    Output.print_comparison_startup()
-
-    # provided printout
-    provided_digest = HashCheck.process_digest(verify_args.digest)
-    print(" Provided :{}".format(provided_digest))
-
-    # stdout used to provide status message while digest is being generated
-    sys.stdout.write(' Generated: {}'.format('Calculating'.center(60)))
-    generated_digest = HashCheck.generate_digest(args.binary)
-    sys.stdout.write("\r Generated:{}\n".format(generated_digest))
-
-    # Compare and printout results
-    result = HashCheck.compare_digests(provided_digest, generated_digest)
-    Output.print_comparison_results(result)
-
-
-if __name__ == '__main__':
-    os.system('cls')
-    args = HashChkParser().get_parser_args()
-    compare_verify_digests(args)
